+++ src/main/java/com/xsushirollx/sushibyte/orderservice/dao/DeliveryDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.xsushirollx.sushibyte.orderservice.model.Delivery;
 
 @Repository
-public interface DeliveryDAO extends JpaRepository<Delivery, Integer> {
+public interface DeliveryDAO extends JpaRepository<Delivery, Long> {
 
 }
